# packages/proto/openapi.yaml
openapi: 3.1.0
info:
  title: Exponential Headless API
  version: 0.1.0
...
servers:
  - url: http://localhost:7016/v1
security:
  - bearerAuth: []
paths:
...
  /customer-requests/{requestID}/projects:
    parameters:
      - name: requestID
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: linkCustomerRequestProject
      summary: Link a customer request to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId]
              properties:
                projectId: { type: string }
      responses:
        '200':
          description: Customer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerRequest'
        default:
          $ref: '#/components/responses/Problem'
  /customer-requests/{requestID}/projects/{projectID}:
    parameters:
      - name: requestID
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: projectID
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: unlinkCustomerRequestProject
      summary: Unlink a customer request from a project
      responses:
        '200':
          description: Customer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerRequest'
        default:
          $ref: '#/components/responses/Problem'
...
  /issues:
    get:
      operationId: listIssues
      summary: List issues in the authenticated workspace
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: team_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Issues page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createIssue
      summary: Create an issue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Created issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        default:
          $ref: '#/components/responses/Problem'
  /issues/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getIssue
      responses:
        '200':
          description: Issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateIssue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
      responses:
        '200':
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteIssue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
        default:
          $ref: '#/components/responses/Problem'
...
  /projects:
    get:
      operationId: listProjects
      summary: List projects in the authenticated workspace
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
...
  /projects/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getProject
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteProject
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: lin_api or pat
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    Problem:
      description: RFC 7807 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    ProjectStatus:
      type: string
      enum: [planned, started, paused, completed, canceled]
    ProjectPriority:
      type: string
      enum: [none, urgent, high, medium, low]
    ProjectTeam:
      type: object
      required: [id, key, name]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
    ProjectProgress:
      type: object
      required: [total, completed, percentage]
      properties:
        total: { type: integer }
        completed: { type: integer }
        percentage: { type: integer }
    Project:
      type: object
      required: [id, name, slug, status, priority, workspace_id, teams, progress, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'] }
        slug: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        priority: { $ref: '#/components/schemas/ProjectPriority' }
        lead_id: { type: [string, 'null'] }
        workspace_id: { type: string, format: uuid }
        start_date: { type: [string, 'null'], format: date-time }
        target_date: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        canceled_at: { type: [string, 'null'], format: date-time }
        teams:
          type: array
          items: { $ref: '#/components/schemas/ProjectTeam' }
        progress: { $ref: '#/components/schemas/ProjectProgress' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ProjectListResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items: { $ref: '#/components/schemas/Project' }
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'], maxLength: 10 }
        slug: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        priority: { $ref: '#/components/schemas/ProjectPriority' }
        lead_id: { type: [string, 'null'] }
        start_date: { type: [string, 'null'], format: date-time }
        target_date: { type: [string, 'null'], format: date-time }
        team_ids:
          type: array
          items: { type: string, format: uuid }
        team_keys:
          type: array
          items: { type: string }
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'], maxLength: 10 }
        slug: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        priority: { $ref: '#/components/schemas/ProjectPriority' }
        lead_id: { type: [string, 'null'] }
        start_date: { type: [string, 'null'], format: date-time }
        target_date: { type: [string, 'null'], format: date-time }
        team_ids:
          type: array
          items: { type: string, format: uuid }
        team_keys:
          type: array
          items: { type: string }
...
    CustomerSummary:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        domain: { type: [string, 'null'] }
        tier: { type: [string, 'null'] }
        status: { type: [string, 'null'] }
    CustomerRequest:
      type: object
      required: [id, workspaceId, customerId, customer, title, important, linkedIssues, linkedProjects, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        title: { type: string }
        body: { type: [string, 'null'] }
        source: { type: [string, 'null'] }
        sourceUrl: { type: [string, 'null'] }
        externalProvider: { type: [string, 'null'] }
        externalId: { type: [string, 'null'] }
        important: { type: boolean }
        createdByUserId: { type: [string, 'null'] }
        linkedIssues:
          type: array
          items:
            $ref: '#/components/schemas/CustomerLinkedIssue'
        linkedProjects:
          type: array
          items:
            $ref: '#/components/schemas/CustomerLinkedProject'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CustomerLinkedIssue:
      type: object
      required: [id, identifier, title, teamKey]
      properties:
        id: { type: string, format: uuid }
        identifier: { type: string }
        title: { type: string }
        teamKey: { type: string }
    CustomerLinkedProject:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    IssuePriority:
      type: string
      enum: [none, urgent, high, medium, low]
    FigmaSource:
      type: object
      required: [id, url, normalizedUrl, fileKey, kind, containerType, capturedAt]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        normalizedUrl: { type: string, format: uri }
        fileKey: { type: string }
        nodeId: { type: [string, 'null'] }
        kind: { type: string, enum: [file, design, proto] }
        name: { type: [string, 'null'] }
        thumbnailUrl: { type: [string, 'null'], format: uri }
        containerType: { type: string, enum: [issue_description, comment, document, plugin] }
        capturedAt: { type: string, format: date-time }
        refreshedAt: { type: [string, 'null'], format: date-time }
        lastError: { type: [string, 'null'] }
    IssueExternalSource:
      type: object
      required: [provider, label, url, externalId, project, integrationId]
      properties:
        provider: { type: string }
        label: { type: string }
        url: { type: string, format: uri }
        externalId: { type: string }
        project: { type: string }
        integrationId: { type: string, format: uuid }
    Issue:
      type: object
      required: [id, number, identifier, title, team_id, state_id, creator_id, priority, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        number: { type: integer }
        identifier: { type: string }
        title: { type: string }
        description: { type: [string, 'null'] }
        team_id: { type: string, format: uuid }
        state_id: { type: string, format: uuid }
        assignee_id: { type: [string, 'null'] }
        creator_id: { type: string }
        priority: { $ref: '#/components/schemas/IssuePriority' }
        estimate: { type: [number, 'null'] }
        parent_issue_id: { type: [string, 'null'], format: uuid }
        project_id: { type: [string, 'null'], format: uuid }
        project_milestone_id: { type: [string, 'null'], format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        due_date: { type: [string, 'null'], format: date-time }
        sort_order: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: [string, 'null'], format: date-time }
        canceled_at: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        figmaSources:
          type: array
          items: { $ref: '#/components/schemas/FigmaSource' }
        sources:
          type: array
          items: { $ref: '#/components/schemas/IssueExternalSource' }
    IssueListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
        next_cursor: { type: [string, 'null'] }
...
    CreateIssueRequest:
      type: object
      required: [title, team_id]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, 'null'] }
        team_id: { type: string, format: uuid }
        state_id: { type: [string, 'null'], format: uuid }
        priority: { $ref: '#/components/schemas/IssuePriority' }
        assignee_id: { type: [string, 'null'] }
        project_id: { type: [string, 'null'], format: uuid }
        project_milestone_id: { type: [string, 'null'], format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        parent_issue_id: { type: [string, 'null'], format: uuid }
        estimate: { type: [number, 'null'] }
        due_date: { type: [string, 'null'], format: date }
    UpdateIssueRequest:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, 'null'] }
        state_id: { type: string, format: uuid }
        priority: { $ref: '#/components/schemas/IssuePriority' }
        assignee_id: { type: [string, 'null'] }
        project_id: { type: [string, 'null'], format: uuid }
        project_milestone_id: { type: [string, 'null'], format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        parent_issue_id: { type: [string, 'null'], format: uuid }
        estimate: { type: [number, 'null'] }
        due_date: { type: [string, 'null'], format: date }
        sort_order: { type: number }
        archive: { type: boolean }